MAILBOX-310 Mailbox query refactoring: Create dedicated packages for Mailbox search related classes
