+++ mailbox/api/src/main/java/org/apache/james/mailbox/model/search/Wildcard.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailbox.model;
+package org.apache.james.mailbox.model.search;
 
 import com.google.common.base.Preconditions;
