+++ mailbox/api/src/test/java/org/apache/james/mailbox/MailboxManagerTest.java
@@ -39,7 +39,7 @@ import org.apache.james.mailbox.model.MailboxAnnotationKey;
 import org.apache.james.mailbox.model.MailboxId;
 import org.apache.james.mailbox.model.MailboxMetaData;
 import org.apache.james.mailbox.model.MailboxPath;
-import org.apache.james.mailbox.model.MailboxQuery;
+import org.apache.james.mailbox.model.search.MailboxQuery;
 import org.junit.Assume;
 import org.junit.Rule;
 import org.junit.Test;
